Extract bytecode manipulation in tests to base class
